feat: VoucherService 인터페이스 생성
1. 테스트용 Service 확장 대비
2. 구현체를 독립적으로 확장할 때 클라이언트의 코드에 영향 주지 않기 위함
