+++ src/main/java/com/programmers/application/service/VoucherService.java
@@ -1,34 +1,13 @@
 package com.programmers.application.service;
 
-import com.programmers.application.domain.voucher.Voucher;
-import com.programmers.application.domain.voucher.VoucherFactory;
-import com.programmers.application.dto.reponse.ResponseFactory;
 import com.programmers.application.dto.reponse.VoucherInfoResponse;
 import com.programmers.application.dto.request.VoucherCreationRequest;
-import com.programmers.application.repository.VoucherRepository;
-import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.UUID;
 
-@Service
-public class VoucherService {
-    private final VoucherRepository voucherRepository;
+public interface VoucherService {
+    UUID createVoucher(VoucherCreationRequest voucherCreationRequest);
 
-    public VoucherService(VoucherRepository voucherRepository) {
-        this.voucherRepository = voucherRepository;
-    }
-
-    public UUID createVoucher(VoucherCreationRequest voucherCreationRequest) {
-        Voucher voucher = VoucherFactory.createVoucher(voucherCreationRequest);
-        voucherRepository.save(voucher);
-        return voucher.getVoucherId();
-    }
-
-    public List<VoucherInfoResponse> findVoucherList() {
-        List<Voucher> voucherList = voucherRepository.findAll();
-        return voucherList.stream()
-                .map(ResponseFactory::createVoucherInfoResponse)
-                .toList();
-    }
+    List<VoucherInfoResponse> findVoucherList();
 }
